Update acf_gearbox.lua
Not sure what happened and why this works now, but I was testing today and it worked perfectly fine, this is very RPM dependent and gear ratios need to be tuned very precisely. Also, it seems that adjusting the gear/final drive on the double diff itself isnt the best option.